Add fake S-Meter sort of thing, narrow mode
Add a fake S Meter.  It's not correct, but I'm a bit unclear on how to really calculate an S-Meter anyway.  But it goes from -66 as noise floor and every 3dB shows a bump.  It's useful just to see how strong the signal is without reading the specific values.

On phones I've found the screen a bit wide, so add a narrow parameter.  Just by itself narrows a bit, add a value also changes the maximum channel name width to what's specified to make it even more narrow.

Also fixes a bug with a regex that wasn't accepting multiple words as a pattern, such as "Range Test" was just evaluating "Range".