Clean up not needed UTF8 stuff, reset monitoring mode once a minute
Previously in monitor mode it would stick on the frequency forever once tuned to make it faster to cycle, however, if something got tweaked it'd be off forever, so reset monitoring once a minute so it re-tunes.

Add tunefailures counter to status bar to see if maybe that's why I see random 0's for the signal level.